kursach-master/src/test/java/com/example/InventoryApp/User/UserControllerTest.java
package com.example.InventoryApp.User;

import com.example.InventoryApp.Controllers.UserController;
import com.example.InventoryApp.Model.User;
import com.example.InventoryApp.Model.UserDTO;
import com.example.InventoryApp.Service.UserService;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;
import org.mockito.InjectMocks;
import org.mockito.Mock;
import org.mockito.MockitoAnnotations;
import org.springframework.http.MediaType;
import org.springframework.test.web.servlet.MockMvc;
import org.springframework.test.web.servlet.setup.MockMvcBuilders;

import java.util.Arrays;
import java.util.List;

import static org.mockito.ArgumentMatchers.any;
import static org.mockito.ArgumentMatchers.eq;
import static org.mockito.Mockito.when;
import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.*;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.*;

import com.fasterxml.jackson.databind.ObjectMapper;

public class UserControllerTest {

    @Mock
    private UserService userService;

    @InjectMocks
    private UserController userController;

    private MockMvc mockMvc;
    private ObjectMapper objectMapper;

    @BeforeEach
    public void setup() {
        MockitoAnnotations.openMocks(this);
        mockMvc = MockMvcBuilders.standaloneSetup(userController).build();
        objectMapper = new ObjectMapper();
    }

    @Test
    public void testGetAll() throws Exception {
        List<User> users = Arrays.asList(
                new User(1L, "User1", "Password1"),
                new User(2L, "User2", "Password2")
        );

        when(userService.getAll()).thenReturn(users);

        mockMvc.perform(get("/api/users"))
                .andExpect(status().isOk())
                .andExpect(content().contentType(MediaType.APPLICATION_JSON))
                .andExpect(jsonPath("$[0].userName").value("User1"))
                .andExpect(jsonPath("$[1].userName").value("User2"));
    }

    @Test
    public void testCreateUser() throws Exception {
        UserDTO newUserDTO = new UserDTO("NewUser", "NewPassword");
        User newUser = new User(1L, "NewUser", "NewPassword");

        when(userService.createUser(any(UserDTO.class))).thenReturn(newUser);

        mockMvc.perform(post("/api/users")
                        .contentType(MediaType.APPLICATION_JSON)
                        .content(objectMapper.writeValueAsString(newUserDTO)))
                .andExpect(status().isOk())
                .andExpect(content().contentType(MediaType.APPLICATION_JSON))
                .andExpect(jsonPath("$.userName").value("NewUser"))
                .andExpect(jsonPath("$.password").value("NewPassword"));
    }

    @Test
    public void testUpdateUser() throws Exception {
        User updatedUser = new User(1L, "UpdatedUser", "UpdatedPassword");

        when(userService.updateUser(eq(1L), any(User.class))).thenReturn(updatedUser);

        mockMvc.perform(put("/api/users/1")
                        .contentType(MediaType.APPLICATION_JSON)
                        .content(objectMapper.writeValueAsString(updatedUser)))
                .andExpect(status().isOk())
                .andExpect(content().contentType(MediaType.APPLICATION_JSON))
                .andExpect(jsonPath("$.userName").value("UpdatedUser"))
                .andExpect(jsonPath("$.password").value("UpdatedPassword"));
    }

    @Test
    public void testDeleteUser() throws Exception {
        mockMvc.perform(delete("/api/users/1"))
                .andExpect(status().isNoContent());
    }
}

